lyricgen/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
openai-whisper
moviepy==1.0.3
librosa==0.10.2
Pillow==11.1.0
numpy<2
google-genai>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
requests>=2.28.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.0
# --- SaaS infrastructure ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
stripe>=8.0.0
# --- Queue, observability, storage ---
redis==5.0.8
rq==1.16.2
lyricsgenius==3.0.1
boto3==1.35.0
sentry-sdk[fastapi]==2.14.0
slowapi==0.1.9
